+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### BREAKING CHANGES
 
 - `apirouter.NewGorillaMuxRouter` is now `gorilla.NewRouter` (exposed by package `github.com/davidebianchi/gswagger/support/gorilla`).
+- removed `apirouter.HandlerFunc`. Now it is exposed by `gorilla.HandlerFunc`
+- changed `apirouter.Router` interface:
+  - now it accept a generics `HandlerFunc` to define the handler function
+  - add method `SwaggerHandler(contentType string, json []byte) HandlerFunc`
+- `NewRouter` function now accept `HandlerFunc` as generics
+
+### Feature
+
+- support to different types of routers
 
 ## 0.6.1 - 17-11-2022